refactor: rdb.tsをrdbディレクトリに分割し責務ごとにファイル化
rdb.tsに集約されていた複数の責務を単一責務のファイルへ分離:
- client.ts: RdbClientの生成（getRdbClient/RdbClient型）
- libsql-node-client.ts: Node向けlibsqlクライアントの動的import
- error.ts: DB呼び出しのResultラップ（wrapDbCall）
- logger.ts: Drizzleクエリロガー
- id.ts: DB ID変換（旧rdb-id.ts）
- index.ts: 公開API再エクスポート

rdb-id.tsもrdbディレクトリ配下へ移動し、インポートパスを更新。
