+++ application/src/domain/article/infrastructure/mapper.ts
@@ -1,6 +1,6 @@
 import type { Article } from '@/domain/article/schema/article-schema'
 import type { Article as RdbArticle } from '@/infrastructure/drizzle-orm/schema'
-import { fromDbId } from '@/infrastructure/rdb-id'
+import { fromDbId } from '@/infrastructure/rdb/id'
 
 export default function fromRdbToArticle(rdbArticle: RdbArticle): Article {
   return {
