+++ application/src/domain/user/infrastructure/mapper.ts
@@ -1,5 +1,5 @@
 import type { ActiveUser as RdbActiveUser } from '@/infrastructure/drizzle-orm/schema'
-import { fromDbId } from '@/infrastructure/rdb-id'
+import { fromDbId } from '@/infrastructure/rdb/id'
 import type { CurrentUser } from '../schema/active-user-schema'
 
 export function mapToActiveUser(activeUser: RdbActiveUser): CurrentUser {
