+++ application/src/domain/user/infrastructure/query-impl.ts
@@ -4,7 +4,7 @@ import { ServerError } from '@/common/errors'
 import { Nullable } from '@/common/types/utility'
 import { activeUsers } from '@/infrastructure/drizzle-orm/schema'
 import { RdbClient, wrapDbCall } from '@/infrastructure/rdb'
-import { toDbId } from '@/infrastructure/rdb-id'
+import { toDbId } from '@/infrastructure/rdb/id'
 import { Query } from '../repository'
 import type { CurrentUser } from '../schema/active-user-schema'
 import { mapToActiveUser } from './mapper'
