+++ application/src/infrastructure/rdb/index.ts
@@ -1,0 +1,2 @@
+export { default, type RdbClient } from '@/infrastructure/rdb/client'
+export { wrapDbCall } from '@/infrastructure/rdb/error'
