+++ application/src/test/helper/user.ts
@@ -2,7 +2,7 @@ import { createClient, type SupabaseClient } from '@supabase/supabase-js'
 import { eq, inArray, like } from 'drizzle-orm'
 import type { ActiveUser } from '@/domain/user/schema/active-user-schema'
 import { activeUsers, users } from '@/infrastructure/drizzle-orm/schema'
-import { fromDbId, toDbIds } from '@/infrastructure/rdb-id'
+import { fromDbId, toDbIds } from '@/infrastructure/rdb/id'
 import TEST_ENV from '@/test/env'
 import app from '@/web/server'
 import { getTestRdb } from './rdb'
